Add phy_print to util.
