+++ examples/InitMatrix.py
@@ -1,27 +1,6 @@
 import phylanx
 et = phylanx.execution_tree
-
-def phy_print(m):
-    ndim = m.num_dimensions()
-    if ndim == 1:
-        for i in range(m.dimension(0)):
-            print(m.get(i))
-    elif ndim == 2:
-        for i in range(m.dimension(0)):
-            for j in range(m.dimension(1)):
-                print("%10.2f" % m.get(i, j), end=" ")
-                if j > 5:
-                    print("...", end=" ")
-                    break
-            print()
-            if i > 5:
-                print("%10s" % "...")
-                break
-    elif ndim == 0:
-        print(m.get(0))
-    else:
-        print("ndim=", ndim)
-
+from phylanx.util import *
 
 # Create a vector of zeros
 vz = et.zeros(10)
